# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: WES Forecast Agentic API
  description: >
    API для Agentic AI-системы прогнозирования почасовой выработки ВЭС.
    Отдаёт координаты турбин, прогнозы с интервалами неопределённости,
    ревизии прогноза, метрики качества, лог шагов агента и чат с агентом.


    **Время.** Все метки времени — UTC (ISO-8601 с `Z`).


    **Прогноз (v1.2).** Прогноз выдаёт ML-модель. Выпуск «как если бы» в день D в 12:00
    местного времени станции (Asia/Almaty, UTC+5) покрывает сутки D+1 и D+2 по местному
    времени — 48 часов. Даты выпуска в backtest: 2026-01-31…2026-02-26 (покрывают 01–28.02).
    Параметры — `GET /api/meta`. Факта за февраль нет в данных: `actualPower` = null.


    **Версии прогноза.** Каждые сутки X прогнозируются дважды: из выпуска X−2 (как D+2) и из
    выпуска X−1 (как D+1, по более свежему прогнозу погоды) — повторный расчёт при
    обновлении входных данных. `GET /api/forecast/revisions`.


    **Честность данных.** Каждый прогноз указывает `weatherIssuedAt` — самый поздний момент
    выпуска использованных прогнозов погоды. Он всегда не позже `forecastIssuedAt`.


    **Метрики** — на отложенном тесте (январь 2026): факта за февраль нет.


    **Цикл агента** — `POST /api/forecast/run`, журнал — `GET /api/agent-log`. Результаты циклов
    сохраняются; отчёт LLM из цикла подставляется в `/api/forecast`.
  version: 1.2.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
...
paths:
...
  /api/forecast:
    get:
      tags: [Forecast]
      summary: Прогноз ML, выпущенный в указанный день
      description: >
        Прогноз, выпущенный «как если бы» в день `date` в 12:00 местного времени
        (`forecastIssuedAt`): 24 ч — следующие сутки, 48 ч — следующие двое суток (по местному
        времени). Кроме точек содержит сводку, предупреждения (ML + правила) и отчёт.
        503 — ML-сервис недоступен или нет погоды на эту дату.
      operationId: getForecast
      parameters:
        - $ref: '#/components/parameters/TurbineId'
        - name: date
          in: query
          required: true
          description: День выпуска (имитация «сегодня»)
          schema:
            type: string
            format: date
          example: "2026-01-31"
        - name: horizonHours
          in: query
          required: false
          description: Горизонт прогноза в часах
          schema:
            type: integer
            enum: [24, 48]
            default: 48
      responses:
        '200':
          description: Прогноз на указанный период
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastResponse'
              example:
                turbineId: "t1"
                forecastIssuedAt: "2026-01-31T07:00:00Z"
                revision: 1
                horizonHours: 48
                weatherSource: "open-meteo-previous-runs"
                weatherIssuedAt: "2026-01-30T18:00:00Z"
                modelVersion: "windml-2026-09-23"
                summary:
                  meanPower: 0.366
                  minPower: 0.0
                  maxPower: 0.901
                  maxPowerAt: "2026-02-02T09:00:00Z"
                  fullLoadHours: 17.6
                  lowPowerHours: 0
                alerts:
                  - type: "low_confidence"
                    severity: "warning"
                    from: "2026-01-31T19:00:00Z"
                    to: "2026-02-02T18:00:00Z"
                    message: "29 ч с разбросом ветра между моделями > 2.5 м/с — рекомендуется пересчёт после следующего обновления прогноза погоды"
                agentReport: "Средняя мощность 0.366, пик 0.901 в 02.02 09:00 UTC, часов с выработкой ниже 0.05: 0."
                points:
                  - timestamp: "2026-01-31T19:00:00Z"
                    predictedPower: 0.098
                    p10: 0.007
                    p90: 0.183
                    actualPower: null
                    windSpeed: 4.1
                    temperature: -6.3
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/Unavailable'
  /api/forecast/revisions:
    get:
      tags: [Forecast]
      summary: Версии прогноза на целевые сутки
      description: >
        Как менялся прогноз на сутки `date` (местное время): версия 1 — из выпуска
        date−2 (как D+2), версия 2 — из выпуска date−1 (как D+1, по более свежей погоде).
        Недоступная версия пропускается. `meanPower` — средняя мощность за эти сутки.
      operationId: getForecastRevisions
      parameters:
        - $ref: '#/components/parameters/TurbineId'
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2026-02-01"
      responses:
        '200':
          description: Список ревизий
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForecastRevisionsResponse'
              example:
                turbineId: "t2"
                date: "2026-02-15"
                revisions:
                  - revision: 1
                    forecastIssuedAt: "2026-02-13T07:00:00Z"
                    weatherIssuedAt: "2026-02-12T18:00:00Z"
                    meanPower: 0.511
                    changeVsPreviousPct: null
                  - revision: 2
                    forecastIssuedAt: "2026-02-14T07:00:00Z"
                    weatherIssuedAt: "2026-02-13T18:00:00Z"
                    meanPower: 0.555
                    changeVsPreviousPct: 8.6
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/forecast/run:
    post:
      tags: [Forecast]
      summary: Запустить цикл агента для турбины и дня выпуска
      description: >
        Цикл агента: погода у турбины (fetchWeather) → прогноз ML с повторами (predict) →
        проверка (validate) → сравнение с предыдущим выпуском (compareWithPrevious) →
        отчёт LLM с самопроверкой (llm) → сохранение (saveForecast). Асинхронный; прогресс —
        поллингом `GET /api/agent-log?date=…&turbineId=…` раз в 1–2 с, пока `status` = running.
        Если цикл для этой пары уже идёт — возвращается его `cycleId`. Обычно занимает 5–15 с.
      operationId: runForecastCycle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunForecastRequest'
      responses:
        '202':
          description: Цикл запущен, обработка асинхронная
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunForecastResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/agent-log:
    get:
      tags: [AgentLog]
      summary: Журнал последнего цикла агента для турбины и дня выпуска
      description: >
        Шаги агента с их статусами и подробностями. У шагов fetchWeather и predict
        `dataIssuedAt` — время выпуска использованных данных погоды; оно не позже
        `forecastIssuedAt` (отметка «✓ данные не из будущего»). Повторная попытка ML —
        отдельный шаг со статусом retrying. 404 — цикл для этой пары ещё не запускался.
      operationId: getAgentLog
      parameters:
        - name: date
          in: query
          required: true
          description: День выпуска
          schema:
            type: string
            format: date
          example: "2026-02-10"
        - name: turbineId
          in: query
          required: false
          description: По умолчанию t1
          schema:
            type: string
      responses:
        '200':
          description: Лог шагов агента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentLogResponse'
              example:
                date: "2026-02-10"
                cycleId: "cycle_t1_2026-02-10_1790165347"
                revision: 1
                forecastIssuedAt: "2026-02-10T07:00:00Z"
                turbineId: "t1"
                status: "success"
                reportSource: "llm"
                steps:
                  - stepName: "Погода у турбины"
                    status: "success"
                    timestamp: "2026-09-23T12:09:07Z"
                    details: "Open-Meteo ecmwf_ifs, прогон от 10.02 05:00, опубликован к 10.02 12:00 — не позже момента прогноза 10.02 12:00"
                    tool: "fetchWeather"
                    dataIssuedAt: "2026-02-10T00:00:00Z"
                  - stepName: "Прогноз ML"
                    status: "success"
                    timestamp: "2026-09-23T12:09:08Z"
                    details: "Модель windml-2026-09-23: 48 ч, погода open-meteo-previous-runs выпущена не позже 09.02 23:00"
                    tool: "predict"
                    dataIssuedAt: "2026-02-09T18:00:00Z"
                  - stepName: "Сравнение с предыдущим выпуском"
                    status: "success"
                    timestamp: "2026-09-23T12:09:09Z"
                    details: "Сутки 11.02: было 59 % номинала (выпуск 2026-02-09), стало 61 % (+3 %) — изменение небольшое, прогноз подтверждён."
                    tool: "compareWithPrevious"
                    dataIssuedAt: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/weather:
    get:
      tags: [Weather]
      summary: Прогноз погоды, известный в заданный момент
      description: >
        Берёт последний прогон модели погоды (по умолчанию ECMWF IFS HRES 9 км,
        Open-Meteo Single Runs API), опубликованный не позже `at`, и отдаёт часы
        (`at`, `at` + horizonHours]. Прогон считается опубликованным через
        `publication-delay` (7 ч) после инициализации. Если прогон недоступен или
        неполный — берётся более старый, пропущенные перечислены в `skippedRuns`.
        Ответы кэшируются в `data/weather-cache/`.
      operationId: getWeather
      parameters:
        - $ref: '#/components/parameters/TurbineId'
        - name: at
          in: query
          required: true
          description: Момент прогноза, UTC
          schema:
            type: string
            format: date-time
          example: "2026-02-01T00:00:00Z"
        - name: horizonHours
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 72
            default: 48
      responses:
        '200':
          description: Почасовой прогноз погоды
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeatherResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          description: Нет доступного прогона (Open-Meteo недоступен, прогона нет в архиве или offline без кэша)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    TurbineId:
      name: turbineId
      in: query
      required: true
      schema:
        type: string
      example: "t1"
  responses:
    BadRequest:
      description: Некорректные параметры
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unavailable:
      description: ML-сервис недоступен или нет данных на эту дату
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Турбина не найдена
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ForecastPoint:
      type: object
      required: [timestamp, predictedPower]
      properties:
        timestamp:
          type: string
          format: date-time
          description: Начало часа, к которому относится значение, UTC
          example: "2026-01-31T19:00:00Z"
        predictedPower:
          type: number
          format: double
          description: Основной прогноз ML (среднее) нормализованной мощности 0..1
          example: 0.42
        p10:
          type: number
          format: double
          nullable: true
          description: Нижняя граница интервала (10-й перцентиль)
          example: 0.30
        p90:
          type: number
          format: double
          nullable: true
          description: Верхняя граница интервала (90-й перцентиль)
          example: 0.55
        actualPower:
          type: number
          format: double
          nullable: true
          description: Фактическая мощность, если известна. За февраль 2026 факта нет — null
          example: null
        windSpeed:
          type: number
          format: double
          nullable: true
          description: Прогноз ветра на 100 м, м/с, известный в момент выпуска; null — нет данных
          example: 7.2
        temperature:
          type: number
          format: double
          nullable: true
          description: Прогноз температуры, °C; null — нет данных
          example: -3.1
    ForecastSummary:
      type: object
      description: Готовые итоги по точкам прогноза (P50)
      required: [meanPower, minPower, maxPower, maxPowerAt, fullLoadHours, lowPowerHours]
      properties:
        meanPower:
          type: number
          format: double
          example: 0.412
        minPower:
          type: number
          format: double
          example: 0.0
        maxPower:
          type: number
          format: double
          example: 0.91
        maxPowerAt:
          type: string
          format: date-time
          example: "2026-02-01T14:00:00Z"
        fullLoadHours:
          type: number
          format: double
          description: >
            Сумма нормализованной мощности по часам = эквивалент часов работы на
            номинале. Умножить на установленную мощность — получится МВт·ч.
          example: 19.8
        lowPowerHours:
          type: integer
          description: Число часов с мощностью ниже 0.05
          example: 6
    ForecastAlert:
      type: object
      required: [type, severity, from, to, message]
      properties:
        type:
          type: string
          enum: [icing, storm_cutout, ramp_down, ramp_up, low_confidence, model_physics_gap]
          description: >
            icing — риск обледенения; storm_cutout — ветер у порога штормового
            отключения; ramp_down/ramp_up — резкое изменение выработки;
            low_confidence — модели погоды расходятся или широкий интервал P10–P90;
            model_physics_gap — ML сильно расходится с физической кривой мощности
        severity:
          type: string
          enum: [info, warning, critical]
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
          description: Последний час интервала включительно
        message:
          type: string
    ForecastResponse:
      type: object
      required: [turbineId, forecastIssuedAt, revision, horizonHours, weatherSource, weatherIssuedAt, modelVersion, summary, alerts, points]
      properties:
        turbineId:
          type: string
          example: "t1"
        forecastIssuedAt:
          type: string
          format: date-time
          description: Момент выпуска «как если бы сегодня» — день D 12:00 местного, в UTC
          example: "2026-01-31T07:00:00Z"
        revision:
          type: integer
          description: Всегда 1 (версии одних суток — в /api/forecast/revisions)
          example: 1
        horizonHours:
          type: integer
          example: 48
        weatherSource:
          type: string
          description: Источник и модель прогноза погоды
          example: "open-meteo-previous-runs"
        weatherIssuedAt:
          type: string
          format: date-time
          description: Самый поздний момент выпуска использованных прогнозов погоды; всегда ≤ forecastIssuedAt
          example: "2026-01-30T18:00:00Z"
        modelVersion:
          type: string
          example: "windml-2026-09-23"
        summary:
          $ref: '#/components/schemas/ForecastSummary'
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/ForecastAlert'
        agentReport:
          type: string
          nullable: true
          description: Текстовый отчёт агента для диспетчера
        points:
          type: array
          items:
            $ref: '#/components/schemas/ForecastPoint'
    ForecastRevision:
      type: object
      required: [revision, forecastIssuedAt, weatherIssuedAt, meanPower]
      properties:
        revision:
          type: integer
          example: 2
        forecastIssuedAt:
          type: string
          format: date-time
        weatherIssuedAt:
          type: string
          format: date-time
        meanPower:
          type: number
          format: double
          description: Средняя мощность за целевые сутки
        changeVsPreviousPct:
          type: number
          format: double
          nullable: true
          description: Изменение средней мощности относительно предыдущей ревизии, %
    ForecastRevisionsResponse:
      type: object
      required: [turbineId, date, revisions]
      properties:
        turbineId:
          type: string
        date:
          type: string
          format: date
        revisions:
          type: array
          items:
            $ref: '#/components/schemas/ForecastRevision'
    RunForecastRequest:
      type: object
      required: [turbineId, date]
      properties:
        turbineId:
          type: string
          example: "t1"
        date:
          type: string
          format: date
          example: "2026-02-01"
        horizonHours:
          type: integer
          enum: [24, 48]
          default: 48
    RunForecastResponse:
      type: object
      required: [cycleId, status]
      properties:
        cycleId:
          type: string
          example: "cycle_2026-02-01T12:34:00Z"
        status:
          type: string
          example: "processing"
...
    AgentStep:
      type: object
      required: [stepName, status, timestamp]
      properties:
        stepName:
          type: string
          example: "Получение прогноза погоды"
        status:
          type: string
          enum: [success, failed, retrying, running]
          example: "success"
        timestamp:
          type: string
          format: date-time
          example: "2026-02-01T00:00:05Z"
        details:
          type: string
          nullable: true
        tool:
          type: string
          nullable: true
          description: >
            Инструмент агента: fetchWeather, predict, validate, compareWithPrevious, llm, saveForecast
          example: "fetchWeather"
        dataIssuedAt:
          type: string
          format: date-time
          nullable: true
          description: Время выпуска использованных внешних данных (для проверки на утечку)
          example: "2026-01-31T12:00:00Z"
    AgentLogResponse:
      type: object
      required: [date, cycleId, revision, forecastIssuedAt, steps, status]
      properties:
        date:
          type: string
          format: date
          example: "2026-02-01"
        cycleId:
          type: string
          example: "cycle_2026-02-01T00:00:00Z"
        revision:
          type: integer
          example: 1
        forecastIssuedAt:
          type: string
          format: date-time
          example: "2026-02-01T00:00:00Z"
        steps:
          type: array
          items:
            $ref: '#/components/schemas/AgentStep'
        turbineId:
          type: string
          nullable: true
          example: "t1"
        status:
          type: string
          enum: [running, success, failed]
          description: running — цикл идёт (продолжать поллинг); failed — прогноз не принят (см. шаги)
        reportSource:
          type: string
          nullable: true
          enum: [llm, template]
          description: llm — отчёт написала LLM и он прошёл самопроверку; template — шаблон
...
    WeatherHour:
      type: object
      description: Час прогноза погоды; поля совпадают с WeatherHour в contracts/ml-service.openapi.yaml
      required: [timestamp]
      properties:
        timestamp:
          type: string
          format: date-time
        windSpeed10m:
          type: number
          nullable: true
          description: м/с
        windSpeed80m:
          type: number
          nullable: true
          description: м/с
        windSpeed100m:
          type: number
          nullable: true
          description: м/с
        windSpeed120m:
          type: number
          nullable: true
          description: м/с
        windDirection100m:
          type: number
          nullable: true
          description: градусы
        windGusts10m:
          type: number
          nullable: true
          description: м/с
        temperature2m:
          type: number
          nullable: true
          description: °C
        relativeHumidity2m:
          type: number
          nullable: true
          description: "%"
        surfacePressure:
          type: number
          nullable: true
          description: гПа
    WeatherForecast:
      type: object
      required: [model, forecastAt, runInitAt, runAvailableAt, gridLat, gridLon, elevation, source, skippedRuns, hours]
      properties:
        model:
          type: string
          example: "ecmwf_ifs"
        forecastAt:
          type: string
          format: date-time
          example: "2026-02-01T00:00:00Z"
        runInitAt:
          type: string
          format: date-time
          description: Время инициализации использованного прогона (это weatherIssuedAt прогноза мощности)
          example: "2026-01-31T12:00:00Z"
        runAvailableAt:
          type: string
          format: date-time
          description: Когда прогон считается опубликованным; всегда ≤ forecastAt
          example: "2026-01-31T19:00:00Z"
        gridLat:
          type: number
          description: Широта узла сетки модели, ближайшего к турбине
        gridLon:
          type: number
        elevation:
          type: number
          description: Высота узла сетки, м
        source:
          type: string
          enum: [cache, api]
        skippedRuns:
          type: array
          items:
            type: object
            required: [runInitAt, reason]
            properties:
              runInitAt:
                type: string
                format: date-time
              reason:
                type: string
        hours:
          type: array
          items:
            $ref: '#/components/schemas/WeatherHour'
    WeatherResponse:
      type: object
      required: [turbineId, forecast]
      properties:
        turbineId:
          type: string
          example: "t1"
        forecast:
          $ref: '#/components/schemas/WeatherForecast'
    ErrorResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: "Turbine not found: turbineId=nope"